daisytools: Tools for Working with Daisy Simulation Output

Provides tools for reading, transforming, and visualizing log files produced by Daisy simulations. Includes support for importing '.dlf' output files, reshaping depth-dependent and time-series data, and creating static and interactive plots for exploratory analysis and comparison of simulation results.

See also

Author

Maintainer: Silas Ørting silas@di.ku.dk (ORCID)